Address PR 553 review findings
- Ship a self-contained fallback layout and stylesheet in the UI package
- Migrate the authentication pages to Bootstrap 5 idioms
- Widen UploadStatus.ReceivedParts to long
- Guard PutActivitySubscription against a null subscription
- Record why two CodeQL user-controlled-bypass reports are false positives

+++ src/idunno.Bluesky/Video/Model/GetUploadStatusResponse.cs
@@ -9,7 +9,7 @@ internal record GetUploadStatusResponse(
     [property: JsonRequired] string JobId,
     [property: JsonRequired, JsonPropertyName("partSizeBytes")] long PartSize,
     [property: JsonRequired] long PartCount,
-    [property: JsonRequired] IReadOnlyCollection<int> ReceivedParts,
+    [property: JsonRequired] IReadOnlyCollection<long> ReceivedParts,
     [property: JsonRequired] DateTimeOffset ExpiresAt,
     [property: JsonRequired] string State,
     string? CompletedJobId,
